storeliveness: refactor test utils
This commit moves some test definitions to a new `testutil.go` file.

Part of: #125066

Release note: None
