+++ pkg/kv/kvserver/storeliveness/BUILD.bazel
@@ -10,8 +10,8 @@ go_library(
         "requester_state.go",
         "support_manager.go",
         "supporter_state.go",
+        "testutils.go",
         "transport.go",
-        "unreliable_store_liveness_handler.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/storeliveness",
     visibility = ["//visibility:public"],
@@ -66,7 +66,6 @@ go_test(
         "//pkg/util/metric",
         "//pkg/util/netutil",
         "//pkg/util/stop",
-        "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
